+++ docs/Changelog.md
@@ -7,6 +7,7 @@ This page is a brief overview of each version.
 - Use a separate queue for indexing embeddings and summaries, to prevent blocking the main SB indexing thread
 - Refactor to use JSR for most Silverbullet imports, and lots of related changes
 - Reduced bundle size
+- Add support for [space-config](https://silverbullet.md/Space%20Config)
 
 ---
 ## 0.3.2